+++ include/marchingCubeGrid.h
@@ -23,17 +23,17 @@ public:
         double value;
         int points[3];
         int gridIndex;
-        glm::vec3 normal;
+        glm::dvec3 normal;
     };
 
 public:
-    MarchingCubeGrid(const double cubeSize, const glm::vec3 minVolume, const glm::vec3 maxVolume);
+    MarchingCubeGrid(const double cubeSize, const glm::dvec3 minVolume, const glm::dvec3 maxVolume);
     ~MarchingCubeGrid();
 
-    void initializeGrid(const double cubeSize, const glm::vec3 minVolume, const glm::vec3 maxVolume);
+    void initializeGrid(const double cubeSize, const glm::dvec3 minVolume, const glm::dvec3 maxVolume);
     //void computeIsoValues(std::vector<unsigned int> surfaceVertices, double influenceRadius, SpatialGridPoints spatialGrid);
-    void computeIsoValues(const std::vector<glm::vec3> points, double influenceRadius);
-    void triangulate(Mesh& mesh, std::vector<glm::vec3> pointNormals, bool computeNormals);
+    void computeIsoValues(const std::vector<glm::dvec3> points, double influenceRadius);
+    void triangulate(Mesh& mesh, std::vector<glm::dvec3> pointNormals, bool computeNormals);
 
     int getNbVertices() {return _resX*_resY*_resZ;}
 
@@ -46,15 +46,15 @@ private:
     unsigned int _resY;
     unsigned int _resZ;
     double _cubeSize;
-    glm::vec3 _volMin;
-    glm::vec3 _dimensions;
+    glm::dvec3 _volMin;
+    glm::dvec3 _dimensions;
 
     void updateNormals();
     unsigned int getGridIndex(unsigned int xIndex, unsigned int yIndex, unsigned int zIndex);
-    unsigned int getEdgePoint(std::vector<MarchingCubeVertex>, int edgeNo, std::vector<glm::vec3>& points);
+    unsigned int getEdgePoint(std::vector<MarchingCubeVertex>, int edgeNo, std::vector<glm::dvec3>& points);
     unsigned int getIndex(unsigned int gridIndex, int component);
-    glm::vec3 getVertexPosition(unsigned int xIndex, unsigned int yIndex, unsigned int zIndex);
-    CloudVolume getCellsInRadius(const glm::vec3 position, double radius);
+    glm::dvec3 getVertexPosition(unsigned int xIndex, unsigned int yIndex, unsigned int zIndex);
+    CloudVolume getCellsInRadius(const glm::dvec3 position, double radius);
     void setScalarValue(unsigned int xIndex, unsigned int yIndex, unsigned int zIndex, double value);
     bool hasVertexIndexes(std::vector<int> vertexIndexes);
 };
