+++ include/mesh.h
@@ -26,13 +26,13 @@ public:
     ~Mesh();
 
     std::vector<Triangle>& triangles() {return _triangles;}
-    std::vector<glm::vec3>& points() {return _points;}
-    std::vector<glm::vec3>& normals() {return _normals;}
+    std::vector<glm::dvec3>& points() {return _points;}
+    std::vector<glm::dvec3>& normals() {return _normals;}
 
 private:
     std::vector<Triangle> _triangles;
-    std::vector<glm::vec3> _points;
-    std::vector<glm::vec3> _normals;
+    std::vector<glm::dvec3> _points;
+    std::vector<glm::dvec3> _normals;
 };
 
 #endif // MESH_H
